Last details (#38)
* Update README.md. Also add sustentacion/Entrega_1/sustentacion.md file

* Rebase with main

* Documentation about Sonarqube, tests and newman

* Enhance integration_testing.md with image and config details

Added an image and updated the configuration details for the integration testing documentation.

* Add image to quality_analysis.md

Added an image to the quality analysis documentation.

* Add link to GitHub Actions run in quality analysis

Added a link to the GitHub Actions run for quality analysis.

* Add image to testing pipeline documentation

Added an image to enhance the documentation.

* Update testing_pipeline.md with GitHub Actions link

Added a link to the GitHub Actions run for the testing pipeline.